Create Redshift Table Statement

This function generates a SQL CREATE TABLE statement for Redshift based on the provided data frame and parameters.

Usage

rs_create_statement(
  df,
  table_name,
  sortkeys = NULL,
  sortkey_style = "compound",
  distkey = NULL,
  distkey_style = "even",
  compression = TRUE
)

Arguments

df

A data frame to use for generating the CREATE TABLE statement.

table_name

A character string representing the name of the table to be created.

sortkeys

A character vector of column names to be used as sort keys (optional).

sortkey_style

A character string specifying the sort key style, either 'compound' or 'interleaved' (default is 'compound').

distkey

A character string representing the distribution key column (optional).

distkey_style

A character string specifying the distribution style, either 'even' or 'all' (default is 'even').

compression

A logical value indicating whether compression should be applied (default is TRUE).

Value

A character string representing the CREATE TABLE statement for Redshift.