Upload Data to S3 Bucket

This function uploads a data frame to an S3 bucket, splitting it into multiple files if necessary. The function checks if the bucket exists and retries in case of errors during upload.

Usage

uploadToS3(data, bucket, split_files, key, secret, session, region)

Arguments

data

A data frame to upload to S3.

bucket

The name of the S3 bucket.

split_files

Number of files to split the data into for upload.

key

AWS access key for authentication.

secret

AWS secret key for authentication.

session

AWS session token for authentication (optional).

region

AWS region where the bucket is located.

Value

A prefix string used for the uploaded files, or NA if the upload fails.